+++ VisualStudio/fheroes2/sources.props
@@ -181,7 +181,6 @@
     <ClCompile Include="src\fheroes2\kingdom\week.cpp" />
     <ClCompile Include="src\fheroes2\maps\ground.cpp" />
     <ClCompile Include="src\fheroes2\maps\maps.cpp" />
-    <ClCompile Include="src\fheroes2\maps\maps_actions.cpp" />
     <ClCompile Include="src\fheroes2\maps\maps_fileinfo.cpp" />
     <ClCompile Include="src\fheroes2\maps\maps_objects.cpp" />
     <ClCompile Include="src\fheroes2\maps\maps_tiles.cpp" />
@@ -364,7 +363,6 @@
     <ClInclude Include="src\fheroes2\kingdom\week.h" />
     <ClInclude Include="src\fheroes2\maps\ground.h" />
     <ClInclude Include="src\fheroes2\maps\maps.h" />
-    <ClInclude Include="src\fheroes2\maps\maps_actions.h" />
     <ClInclude Include="src\fheroes2\maps\maps_fileinfo.h" />
     <ClInclude Include="src\fheroes2\maps\maps_objects.h" />
     <ClInclude Include="src\fheroes2\maps\maps_tiles.h" />
